requirements.txt
aiogram==3.15.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.9.2
pydantic-settings==2.13.1
apscheduler==3.10.4
openai==1.109.1
aiohttp==3.10.11
python-dateutil==2.9.0.post0
